+++ pinn/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-__version__ = '2.1.0'
+__version__ = '2.1.1'
 
 from pinn.networks import get as get_network
 from pinn.models import get as get_model
@@ -10,7 +10,8 @@ def get_calc(model_spec, **kwargs):
     """Get a calculator from a trained model.
 
     The positional argument will be passed to `pinn.get_model`, keyword
-    arguments will be passed to the calculator.
+    arguments will be passed to the calculator. ``default_dtype`` is a
+    calculator argument (not YAML): ``None`` follows ``settings.dtype``.
     """
     import tensorflow as tf
     from pinn import get_model